requirements.txt
# Core dependencies
streamlit>=1.28.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

# AI and Machine Learning
ultralytics>=8.0.0
pytesseract>=0.3.10
openai>=1.0.0

# Image processing
opencv-python-headless>=4.8.0

# Environment and configuration
python-dotenv>=1.0.0

# Data handling
pandas>=2.0.0

# Utilities
requests>=2.31.0
urllib3>=2.0.0

# Development and testing (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 